Change: Parameter transformation replaces individual parameters
Previously the parameter transformation XML replaced the whole list of parameters that the request was processing. Now the parameter transformation replaces individual parameters. If the parameter is not defined in the endpoints.xml, this is an error. If a parameter is not produced in the output of the XSLT, then it is unaltered.