Split doc between "Endpoints" to "XSLT Library" projects appropriately
Previously all the documentation was in .md files in the "Endpoints"
project, even for code which was implemented in the "XSLT Library"
project. Even though this was easier to read for users of the
"Endpoints" project, it was confusing for those projects which only used
the "XSLT Library", or for creating atomic commits which altered both
code and documentation of those features implemented in the "XSLT
Library" project.